Update flair_helper2.py
Fixed a couple bugs where if flair was being applied to a post by a suspended user, it would error out as it was unable to fetch the id of the author.  This has been fixed, as well as I've implemented a check if a user is suspended or deleted, and to not process bans/unbans/set author text or css flair/usernotes/add or remove contributor.