Update flair_helper2_async.py
Bit of an overhaul, the bot will now first attempt to load the configs from it's local db upon startup (this significantly reduces the time for it to start up and begin monitoring.  For example, with 15 or so subs, at about 3 seconds per sub to fetch and cache wiki pages on startup, that's reduced startup time from 45 seconds down to just a couple seconds.

You can set the wiki_fetch_delay under main() to any other value you want, I have it set to 90 in case there is a small inrush of flairs upon start-up, so after 90 seconds it will then fetch and cache any wikis that may have changed, otherwise if unchanged, it continues to use it's local configuration that was already cached.

Various other small tweaks.