Big Update: Actions Database
This update has a pretty significant change in that actions that are detected from the mod_log stream, are not directly sent to process_flair_assignment to be actioned anymore.

Instead, the actions required, based upon the flair helper wiki configuration, are piped into a separate "actions" database stored along with the submission ID.

From there, there is a new process called process_flair_actions, which is what reads from the actions database in order to send them to process_flair_assignment for processing.

This means that if the bot is API Rate Limited, or if it crashes (for whatever reason), even if it crashes or gets disconnected mid-api call while processing a submission, any of the actions in the actions database are only marked as 'completed' once they have actually been processed.  So this means if you kill the bot while it's in the middle of doing something, or if it crashes.  The next time it's started back up, it will only process what's left in the actions database that hadn't been marked as completed yet.

I noticed occasionally when the reddit servers hiccupped, even with the extensive error handling sometimes it would just outright get disconnected, either due to being heavily used, or if Reddit goes hard-down and you see the whole "you broke reddit".