Update flair_helper2_async.py
2 Major Changes

- Fixed the JSON Validation and detection.  I encountered an issue where a specific malformed json config would cause a looping error.  The corrected code now checks if the wiki begins with '[' which would indicate a JSON config, and then attempt to validate it before continuing.  It now correctly identifies the issue and will send a PM to the user making the change (this can be changed in the code to PM the subreddit instead if you wish for the whole team to be aware of an error, the code is simply commented out on line 483/484.

Secondly, there is a new process for the bot to run, I noticed after several weeks of uptime, the prior version of the script eventually started to duplicate actions as if it were running multiple threads, although I was unable to determine what caused it.  A new bit of code will keep track of the different tasks, and if one needs restarting, it first will check if the task exists, and if so, will kill it before spawning a new thread so hopefully this should prevent duplicate instances of the monitor_mod_log threads from spawning and catching actions twice or three times.