Update flair_helper2_async.py
Fixed an issue where attempting to run certain flair actions on an account that is shadowbanned or has the strike through their username, while the post is still accessible, for some reason occasionally (and randomly) the post is not accessible via the API, so the bot would get stuck on those posts repeatedly and indefinitely until you deleted the actions database file.  It was hard to narrow down.  This update contains a fix where if it detects an error like this while attempting to run through the actions, it marks all the actions as complete, and the function that runs through them (process_flair_actions) sees the submission is complete, it deletes the completed actions so the bot tries once, fails, and purges the entry and moves on.