Update flair_helper2_async.py
Changed quite a number of things.

Added an ability to ban and nuke a users content across multiple subs that your instance of flair helper 2 moderates (be mindful of this if hosting for someone else, as their config will allow flair helper to ban and nuke content from a user across multiple subreddits.  This has to explicitly be enabled in the config with  allow_ban_and_nuke = True

Furthermore, have attempted to enhance the automatic restarting of the flair helper tasks that run to monitor the mod log, pms, and and the function to process flair actions.  Reddit has been increasingly randomly dishing out HTTP Error 400 responses which have previously caused the process to simply fail.  I've continually been trying to build a mechanism to automatically restart the individual processes when it receives said HTTP error, although will continue to 

I've created the ability to set up a Telegram bot for status monitoring, as well as to send a /restart command to restart the processes, or /kill to outright kill the entire python project.  (If you have it running as a Linux service with it set to auto-restart, the /kill command can effectively make it do a full restart of the script, although you will need to set up a telegram bot account with @botfather first.

Other iterative improvements to the actions database (which is what keeps track of pending actions to perform in case the bot becomes rate-limited, or crashes mid-action, the pending actions will remain in the database once the bot's tasks restart, and it will then see there are still pending actions in the database and run through to complete them.)

I'd suggest just viewing the actual code changes to see any other changes that were made between the last stable version and this one.